chore: move funcitons to files; max allowed map size
